requirements.txt
absl-py==0.2.2
aioredis==1.1.0
asgiref==2.3.2
asn1crypto==0.24.0
astor==0.6.2
async-timeout==3.0.0
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
backcall==0.1.0
bleach==1.5.0
certifi==2018.8.24
cffi==1.11.5
channels==2.1.2
channels-redis==2.3.0
chardet==3.0.4
config==0.4.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
cycler==0.10.0
daphne==2.2.2
decorator==4.3.0
dj-database-url==0.5.0
Django==2.1
django-cors-headers==2.4.0
django-filter==1.1.0
django-heroku==0.3.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.3
gast==0.2.0
grpcio==1.12.1
gunicorn==19.9.0
h5py==2.8.0
hiredis==0.2.0
html5lib==0.9999999
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipython==6.4.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.12.0
Jinja2==2.10
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
msgpack==0.5.6
numpy==1.14.4
pandas==0.23.0
parso==0.2.1
patsy==0.5.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
psycopg2-binary==2.7.5
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pyparsing==2.2.0
python-axolotl==0.1.42
python-axolotl-curve25519==0.4.1.post2
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
requests==2.19.1
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8.1
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
statsmodels==0.9.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
traitlets==4.3.2
Twisted==18.7.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
whitenoise==4.0
yowsup2==2.5.2
zope.interface==4.5.0
